a little refactoring of the code of object oriented application implementation
