+++ README.md
@@ -1,3 +1,2 @@
-[![Java CI with Maven](https://github.com/Brest-Java-Course-2021/vPivovar/actions/workflows/maven.yml/badge.svg)](https://github.com/Brest-Java-Course-2021/vPivovar/actions/workflows/maven.yml)
 # vPivovar
-The simple price-calculator
+An object-oriented version of a simple calculator implementation
