+++ price-calc-OOP/src/main/java/com/epam/brest/messaging/MessageRequester.java
@@ -1,6 +1,6 @@
 package com.epam.brest.messaging;
 
-import com.epam.brest.request.Requester;
+import com.epam.brest.requesters.Requester;
 import com.epam.brest.messengers.MessengerOutput;
 
 import java.io.*;
