+++ price-calc-OOP/src/main/java/com/epam/brest/messaging/MessengerBeforeRequester.java
@@ -1,6 +1,6 @@
 package com.epam.brest.messaging;
 
-import com.epam.brest.request.Requester;
+import com.epam.brest.requesters.Requester;
 
 import java.io.OutputStream;
