+++ price-calc-OOP/src/main/java/com/epam/brest/messengers/ErrorMessengerRequesterAbstract.java
@@ -1,7 +1,7 @@
 package com.epam.brest.messengers;
 
-import com.epam.brest.request.Requester;
-import com.epam.brest.request.RequesterAbstract;
+import com.epam.brest.requesters.Requester;
+import com.epam.brest.requesters.RequesterAbstract;
 
 public abstract class ErrorMessengerRequesterAbstract<R, I> extends RequesterAbstract<R, I> {
