+++ price-calc-OOP/src/main/java/com/epam/brest/price/DeliveryPriceCalculator.java
@@ -1,29 +1,27 @@
 package com.epam.brest.price;
 
-import com.epam.brest.price.functions.FunctionFactory;
-import com.epam.brest.request.Requester;
-
-import java.util.function.Function;
+import com.epam.brest.price.factories.requesters.RequesterFactory;
+import com.epam.brest.requesters.Requester;
 
 public class DeliveryPriceCalculator implements Requester<Long> {
 
-    private final Function<Void, Long> pricePerKg;
-    private final Function<Void, Long> pricePerKm;
+    private final Requester<Long> pricePerKg;
+    private final Requester<Long> pricePerKm;
 
-    public DeliveryPriceCalculator(Function<Void, Long> pricePerKg,
-                                   Function<Void, Long> pricePerKm) {
+    public DeliveryPriceCalculator(Requester<Long> pricePerKg,
+                                   Requester<Long> pricePerKm) {
         this.pricePerKg = pricePerKg;
         this.pricePerKm = pricePerKm;
     }
 
-    public DeliveryPriceCalculator(FunctionFactory<Void, Long> pricePerKgFactory,
-                                   FunctionFactory<Void, Long> pricePerKmFactory) {
+    public DeliveryPriceCalculator(RequesterFactory<Long> pricePerKgFactory,
+                                   RequesterFactory<Long> pricePerKmFactory) {
         this.pricePerKg = pricePerKgFactory.create();
         this.pricePerKm = pricePerKmFactory.create();
     }
 
     @Override
     public Long request() {
-        return pricePerKg.apply(null) + pricePerKm.apply(null);
+        return pricePerKg.request() + pricePerKm.request();
     }
 }
