+++ price-calc-OOP/src/main/java/com/epam/brest/price/PriceCalculator.java
@@ -1,12 +1,11 @@
 package com.epam.brest.price;
 
-import com.epam.brest.request.Requester;
+import com.epam.brest.requesters.Requester;
 
 import java.util.List;
 import java.util.SortedMap;
-import java.util.function.Function;
 
-public class PriceCalculator implements Function<Void, Long> {
+public class PriceCalculator implements Requester<Long> {
 
     private final SortedMap<Double, Long> priceCurve;
     private final Requester<Double> requester;
@@ -18,7 +17,7 @@ public class PriceCalculator implements Function<Void, Long> {
     }
 
     @Override
-    public Long apply(Void unused) {
+    public Long request() {
         var aDouble = requester.request();
         var result = 0L;
         var it = priceCurve.entrySet().iterator();
@@ -29,7 +28,8 @@ public class PriceCalculator implements Function<Void, Long> {
             if (aDouble > next.getKey()) {
                 result += Math.round((next.getKey() - current.getKey()) * current.getValue());
             } else {
-                result += Math.round((current.getKey() - aDouble) * current.getValue());
+                result += Math.round((aDouble - current.getKey()) * current.getValue());
+                break;
             }
         }
