+++ price-calc-OOP/src/main/java/com/epam/brest/price/factories/requesters/RequesterFactory.java
@@ -1,7 +1,7 @@
 package com.epam.brest.price.factories.requesters;
 
 import com.epam.brest.price.factories.Factory;
-import com.epam.brest.request.Requester;
+import com.epam.brest.requesters.Requester;
 
 public interface RequesterFactory<R> extends Factory<Requester<R>> {
     Requester<R> create();
