+++ price-calc-OOP/src/main/java/com/epam/brest/requesters/Requester.java
@@ -1,4 +1,4 @@
-package com.epam.brest.request;
+package com.epam.brest.requesters;
 
 public interface Requester<R> {
