+++ price-calc-OOP/src/main/java/com/epam/brest/requesters/RequesterAbstract.java
@@ -1,4 +1,4 @@
-package com.epam.brest.request;
+package com.epam.brest.requesters;
 
 public abstract class RequesterAbstract<R, I> implements Requester<R> {
