+++ price-calc-OOP/src/main/java/com/epam/brest/requesters/WeightFactoryPriceCalculator.java
@@ -1,14 +1,15 @@
-package com.epam.brest.price.functions;
+package com.epam.brest.requesters;
 
 import com.epam.brest.price.PriceCalculator;
+import com.epam.brest.price.factories.requesters.RequesterFactory;
 import com.epam.brest.price.factories.requesters.RequesterFactoryWeightConsole;
+import com.epam.brest.requesters.Requester;
 
 import java.util.TreeMap;
-import java.util.function.Function;
 
-public class WeightFactoryPriceCalculator implements FunctionFactory<Void, Long> {
+public class WeightFactoryPriceCalculator implements RequesterFactory<Long> {
     @Override
-    public Function<Void, Long> create() {
+    public Requester<Long> create() {
         var priceCurve = new TreeMap<Double, Long>();
         priceCurve.put(0D, 1000L);
         priceCurve.put(100D, 2000L);
